feat: migrate dusta apps v1 endpoints to vaults
