+++ front/pages/api/v1/w/[wId]/apps/[aId]/runs/[runId]/index.ts
@@ -1,22 +1,9 @@
-import type { RunType, WithAPIErrorResponse } from "@dust-tt/types";
-import { CoreAPI } from "@dust-tt/types";
-import type { NextApiRequest, NextApiResponse } from "next";
-
-import apiConfig from "@app/lib/api/config";
-import { withPublicAPIAuthentication } from "@app/lib/api/wrappers";
-import type { Authenticator } from "@app/lib/auth";
-import { AppResource } from "@app/lib/resources/app_resource";
-import logger from "@app/logger/logger";
-import { apiError } from "@app/logger/withlogging";
+import handler from "@app/pages/api/v1/w/[wId]/vaults/[vId]/apps/[aId]/runs/[runId]";
 
 export const config = {
   api: {
     responseLimit: "8mb",
   },
-};
-
-export type GetRunResponseBody = {
-  run: RunType;
 };
 
 /**
@@ -60,91 +47,4 @@ export type GetRunResponseBody = {
  *                   $ref: '#/components/schemas/Run'
  */
 
-async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse<WithAPIErrorResponse<GetRunResponseBody>>,
-  auth: Authenticator
-): Promise<void> {
-  const { aId } = req.query;
-  if (typeof aId !== "string") {
-    return apiError(req, res, {
-      status_code: 400,
-      api_error: {
-        type: "invalid_request_error",
-        message: "The app ID is invalid.",
-      },
-    });
-  }
-
-  const owner = auth.getNonNullableWorkspace();
-
-  const app = await AppResource.fetchById(auth, aId);
-  if (!app) {
-    return apiError(req, res, {
-      status_code: 404,
-      api_error: {
-        type: "app_not_found",
-        message: "The app you're trying to run was not found",
-      },
-    });
-  }
-  const coreAPI = new CoreAPI(apiConfig.getCoreAPIConfig(), logger);
-
-  switch (req.method) {
-    case "GET":
-      const runId = req.query.runId as string;
-
-      logger.info(
-        {
-          workspace: {
-            sId: owner.sId,
-            name: owner.name,
-          },
-          app: app.sId,
-          runId,
-        },
-        "App run retrieve"
-      );
-
-      // TODO(spolu): This is borderline security-wise as it allows to recover a full run from the
-      // runId assuming the app is public. We should use getRun and also enforce in getRun that we
-      // retrieve only our own runs. Basically this assumes the `runId` as a secret.
-      const runRes = await coreAPI.getRun({
-        projectId: app.dustAPIProjectId,
-        runId,
-      });
-      if (runRes.isErr()) {
-        return apiError(req, res, {
-          status_code: 400,
-          api_error: {
-            type: "run_error",
-            message: "There was an error retrieving the run.",
-            run_error: runRes.error,
-          },
-        });
-      }
-      const run: RunType = runRes.value.run;
-      run.specification_hash = run.app_hash;
-      delete run.app_hash;
-
-      if (run.status.run === "succeeded" && run.traces.length > 0) {
-        run.results = run.traces[run.traces.length - 1][1];
-      } else {
-        run.results = null;
-      }
-
-      res.status(200).json({ run });
-      return;
-
-    default:
-      return apiError(req, res, {
-        status_code: 405,
-        api_error: {
-          type: "method_not_supported_error",
-          message: "The method passed is not supported, GET is expected.",
-        },
-      });
-  }
-}
-
-export default withPublicAPIAuthentication(handler);
+export default handler;
